requirements.txt
# app requirements
typer[all]
ascii
regex
packaging
html-clean
lxml>=5.4.0,<7.0.0
pyease-grpc>=1.6.0
python-dotenv>=0.15.0,<2.0.0
beautifulsoup4>=4.8.0,<5.0.0
requests>=2.31.0
requests_toolbelt>=1.0.0
websocket-client >= 1.7.0
python-slugify>=4.0.0,<9.0.0
colorama>=0.4.0,<0.5.0
tqdm>=4.60,<5.0
PyExecJS>=1.5.1,<2.0.0
ebooklib>=0.17.0,<1.0.0
pillow>=6.0.0
readability-lxml>=0.8.0,<1.0.0
questionary>=1.6.0
prompt-toolkit~=3.0
html5lib~=1.1
base58~=2.1.1
python-box>=6.0.0,<8.0.0
pycryptodome>=3.0.0,<4.0.0
selenium>=3.141.0
tenacity>=9.0.0

# cloudscraper>=1.2.71
pyparsing >= 3.1.0
pyOpenSSL >= 24.0.0
brotli >= 1.1.0

# message bot requirements
discord.py>=2.0.0
python-telegram-bot[job-queue]~=20.0

# server requirements
uvicorn
fastapi[standard]
cachetools
sqlmodel
psycopg[binary]
passlib[argon2]
python-jose[cryptography]
python-dateutil

# dev requirements
wheel
build
black
flake8
tk-tools
setuptools>=61.0
pyinstaller

# types
types-tqdm
types-colorama
types-cachetools
types-python-dateutil
types-passlib
types-python-jose
types-setuptools
lxml-stubs
